// src/app/cjf/enum/tabela-enum.ts
export enum TabelaEnum {
    ASSUNTO = 'assunto',
    UNIDADE_JUDICIAL = 'unidade-judicial',
    UNIDADE = 'unidade',
    CLASSIFICACAO_DESPESA = 'classificacao-despesa',
    INDICE = 'indice',
    INDICE_TIPO = 'indice-tipo',
    BENEFICIARIO_TIPO = 'beneficiario-tipo',
    BENEFICIARIO_SUCESSAO_TIPO = 'beneficiario-sucessao-tipo',
    BENEFICIARIO_VALOR_ORDEM_PAGAMENTO_107A_TIPO = 'beneficiario-valor-ordem-pagamento-107a-tipo',
    SERVIDOR_CONDICAO_TIPO = 'servidor-condicao-tipo',
    FASE_TIPO = 'fase-tipo',
    IDENTIFICACAO_TIPO = 'identificacao-tipo',
    MOVIMENTO_TIPO = 'movimento-tipo',
    SENTENCA_TIPO = 'sentenca-tipo',
    UNIDADE_JUDICIAL_TIPO = 'unidade-judicial-tipo',
    VALOR_TIPO = 'valor-tipo',
    DIVERGENCIA_TIPO = 'divergencia-tipo'
}